Move `TunConfig` to Talpid package
