+++ android/src/main/kotlin/net/mullvad/talpid/tun_provider/TunConfig.kt
@@ -1,7 +1,6 @@
-package net.mullvad.mullvadvpn.model
+package net.mullvad.talpid.tun_provider
 
 import java.net.InetAddress
-import net.mullvad.talpid.tun_provider.InetNetwork
 
 data class TunConfig(
     val addresses: List<InetAddress>,
