+++ mullvad-jni/src/lib.rs
@@ -60,7 +60,6 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/mullvadvpn/model/RelaySettingsUpdate$CustomTunnelEndpoint",
     "net/mullvad/mullvadvpn/model/RelaySettingsUpdate$RelayConstraintsUpdate",
     "net/mullvad/mullvadvpn/model/Settings",
-    "net/mullvad/mullvadvpn/model/TunConfig",
     "net/mullvad/mullvadvpn/model/TunnelState$Blocked",
     "net/mullvad/mullvadvpn/model/TunnelState$Connected",
     "net/mullvad/mullvadvpn/model/TunnelState$Connecting",
@@ -73,6 +72,7 @@ const CLASSES_TO_LOAD: &[&str] = &[
     "net/mullvad/talpid/net/TransportProtocol$Udp",
     "net/mullvad/talpid/net/TunnelEndpoint",
     "net/mullvad/talpid/tun_provider/InetNetwork",
+    "net/mullvad/talpid/tun_provider/TunConfig",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Block",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Nothing",
     "net/mullvad/talpid/tunnel/ActionAfterDisconnect$Reconnect",
